+++ R/import_as.R
@@ -1,4 +1,4 @@
-#' Load R-package and its Re-exports and/or its (Reverse) Dependencies Under a Single Alias
+#' Import R-package and its Re-exports and/or its (Reverse) Dependencies Under a Single Alias
 #'
 #' @description
 #'
@@ -10,31 +10,31 @@
 #' (like the global environment, or the environment within a function). \cr
 #'
 #' @param alias a syntactically valid non-hidden name giving the alias object
-#' where the package(s) are to be loaded into. \cr
+#' where the package(s) are to be imported into. \cr
 #' This name can be given either as a single string (i.e. \code{"alias."}),
 #' or as a one-sided formula with a single term (i.e. \code{~ alias.}).
 #' @param main_package a single string,
-#' giving the name of the main package to load under the given alias.
+#' giving the name of the main package to import under the given alias.
 #' @param re_exports \code{TRUE} or \code{FALSE}.
 #'  * If \code{re_exports = TRUE} the re-exports from the \code{main_package}
 #'  are added to the alias together with the main package.
 #'  This is the default, as it is analogous to the behaviour of base R's \link{::} operator.
 #'  * If \code{re_exports = FALSE}, these re-exports are not added together with the main package.
-#'  The user can still load the packages from which the re-exported functions came from,
+#'  The user can still import the packages under the alias from which the re-exported functions came from,
 #'  by specifying them in the \code{dependencies} argument.
 #' @param dependencies an optional character vector,
 #' giving the names of the dependencies of the
-#' \code{main_package} to be loaded also under the alias. \cr
-#' Defaults to \code{NULL}, which means no dependencies are loaded. \cr
+#' \code{main_package} to be imported also under the alias. \cr
+#' Defaults to \code{NULL}, which means no dependencies are imported under the alias. \cr
 #' See \link{pkg_get_deps} to quickly get dependencies from a package. \cr
 #' @param extensions an optional character vector,
 #' giving the names of the extensions of the
-#' \code{main_package} to be loaded also under the alias. \cr
-#' Defaults to \code{NULL}, which means no extensions are loaded. \cr
-#' @param loadorder the character vector \cr
+#' \code{main_package} to be imported also under the alias. \cr
+#' Defaults to \code{NULL}, which means no extensions are imported under the alias. \cr
+#' @param import_order the character vector \cr
 #' \code{c("dependencies", "main_package", "extensions")}, \cr
 #' or some re-ordering of this character vector,
-#' giving the relative load order of the groups of packages. \cr
+#' giving the relative import order of the groups of packages. \cr
 #' See Details section for more information. \cr
 #' @param lib.loc character vector specifying library search path
 #' (the location of R library trees to search through). \cr
@@ -73,7 +73,7 @@
 #' This becomes cumbersome as more packages are needed and/or
 #' as the package name(s) become longer. \cr
 #' The \code{import_as()} function avoids this issue
-#' by allowing multiple \bold{related} packages to be loaded under a single alias,
+#' by allowing multiple \bold{related} packages to be imported under a single alias,
 #' allowing one to code like this:
 #'
 #' ```{r eval = FALSE}
@@ -87,9 +87,9 @@
 #' alias.$some_function3()
 #' ```
 #'
-#' Thus loading a package, or multiple directly related packages, under a single alias,
+#' Thus importing a package, or multiple directly related packages, under a single alias,
 #' which \code{import_as()} provides, avoids the above issues.
-#' Loading (a) package(s) under an alias is known as "aliasing" (a) package(s). \cr
+#' Importing (a) package(s) under an alias is known as "aliasing" (a) package(s). \cr
 #' \cr
 #' Notice that the \code{import_as()} function has the \code{lib.loc} argument,
 #' allowing to specify the library path,
@@ -103,16 +103,16 @@
 #' with a dot (\code{.}) or underscore (\code{_}). \cr
 #' \cr
 #' \cr
-#' \bold{Regarding the Load Order} \cr
+#' \bold{Regarding \code{import_order}} \cr
 #' The order of the character vector given in
 #' the \code{dependencies} and \code{extensions} arguments matters.
 #' If multiple packages share objects with the same name,
 #' the objects of the package named last will overwrite those of the earlier named packages. \cr
 #' \cr
-#' The \code{loadorder} argument defaults to the character vector \cr
+#' The \code{import_order} argument defaults to the character vector \cr
 #' \code{c("dependencies", "main_package", "extensions")}, \cr
 #' which is the recommended setting. \cr
-#' This setting results in the following load order: \cr
+#' This setting results in the following importing order: \cr
 #'  1) The dependencies, \bold{in the order specified by the \code{depenencies} argument}.
 #'  2) The main_package (see argument \code{main_package}),
 #' including re-exports (if \code{re_exports = TRUE}).
@@ -123,7 +123,7 @@
 #'
 #' \bold{Other Details} \cr
 #' The \code{import_as()} function
-#' does not support loading base/core R under an alias. \cr
+#' does not support importing base/core R under an alias. \cr
 #' Packages that appear in the "Suggests" or "Enhances" fields of packages
 #' are not considered dependencies or extensions. \cr
 #' \cr
@@ -141,7 +141,7 @@
 #' To use, for example, function "some_function()" from alias "alias.", use: \cr
 #' \code{alias.$some_function()} \cr
 #' To see the special attributes of this alias object, use \link{attr.import}. \cr
-#' To "unload" the package alias object, simply remove it (i.e. \code{rm(list="alias.")}). \cr
+#' To "unimport" the package alias object, simply remove it (i.e. \code{rm(list="alias.")}). \cr
 #'
 #' @seealso [tinycodet_import()]
 #'
@@ -166,7 +166,7 @@ import_as <- function(
     alias, main_package, re_exports = TRUE,
     dependencies = NULL, extensions = NULL,
     lib.loc = .libPaths(),
-    loadorder = c("dependencies", "main_package", "extensions")
+    import_order = c("dependencies", "main_package", "extensions")
 ) {
 
   # Check alias:
@@ -210,9 +210,9 @@ import_as <- function(
     stop("`re_exports` must be either `TRUE` or `FALSE`")
   }
 
-  # check load order:
-  if(!.loadorder_is_correct(loadorder)) {
-    stop("Improper load order given")
+  # check import order:
+  if(!.import_order_is_correct(import_order)) {
+    stop("Improper `import_order` given")
   }
 
   # check dependencies + extensions combo:
@@ -220,7 +220,7 @@ import_as <- function(
     stop("packages cannot be both dependencies and extensions!")
   }
   if((length(dependencies) + length(extensions) + 1) > 10) {
-    stop("more than 10 packages not allowed to be loaded under a single alias")
+    stop("more than 10 packages not allowed to be imported under a single alias")
   }
   
   # Check dependencies:
@@ -246,15 +246,12 @@ import_as <- function(
   pkgs <- list(
     dependencies=dependencies, main_package=main_package, extensions=extensions
   )
-  pkgs <- pkgs[loadorder]
+  pkgs <- pkgs[import_order]
   pkgs <- do.call(c, pkgs)
   pkgs <- unique(pkgs)
-
-  if(length(pkgs)>10) {
-    message("Be careful: More than 10 packages are being loaded under the same alias")
-  }
-
-  # load packages:
+  
+  
+  # import packages:
   export_names_all <- character()
   export_names_allconflicts <- character()
   conflicts_df <- data.frame(
@@ -304,7 +301,7 @@ import_as <- function(
   args <- list(
     main_package = main_package, re_exports = re_exports,
     dependencies = dependencies, extensions = extensions,
-    lib.loc = lib.loc, loadorder = loadorder
+    lib.loc = lib.loc, import_order = import_order
   )
   if(isTRUE(re_exports)){
     re_exports.pkgs <- lapply(foreignexports, \(x)attr(x, "package")) |>
@@ -362,7 +359,7 @@ import_as <- function(
   }
   check_args <- isTRUE(all(names(args) %in% c("main_package", "re_exports",
                                               "dependencies", "extensions",
-                                              "lib.loc", "loadorder")))
+                                              "lib.loc", "import_order")))
   if(!check_args){
     return(FALSE)
   }
@@ -371,7 +368,7 @@ import_as <- function(
     isTRUE(args$re_exports) | isFALSE(args$re_exports),
     isTRUE(is.character(args$dependencies) | is.null(args$dependencies)),
     isTRUE(is.character(args$extensions) | is.null(args$extensions)),
-    isTRUE(.loadorder_is_correct(args$loadorder))
+    isTRUE(.import_order_is_correct(args$import_order))
   )
   if(any(!check_args)) {
     return(FALSE)
@@ -379,7 +376,7 @@ import_as <- function(
   pkgs <- list(
     dependencies=args$dependencies, main_package=args$main_package, extensions=args$extensions
   )
-  pkgs <- pkgs[args$loadorder]
+  pkgs <- pkgs[args$import_order]
   pkgs <- do.call(c, pkgs)
   pkgs <- unique(pkgs)
   check_pkgs <- isTRUE(all(pkgs == obj$.__attributes__.$pkgs$packages_order))
@@ -418,21 +415,21 @@ import_as <- function(
 
 #' @keywords internal
 #' @noRd
-.loadorder_is_correct <- function(loadorder) {
-  if(!is.character(loadorder)) {
+.import_order_is_correct <- function(import_order) {
+  if(!is.character(import_order)) {
     return(FALSE)
   }
-  loadorder <- tolower(loadorder)
-  if(length(loadorder) != 3) {
+  import_order <- tolower(import_order)
+  if(length(import_order) != 3) {
     return(FALSE)
   }
-  if(anyDuplicated(loadorder)) {
+  if(anyDuplicated(import_order)) {
     return(FALSE)
   }
-  check_loadorder <- all(
-    sort(loadorder) ==sort(c("dependencies", "main_package", "extensions"))
+  check_import_order <- all(
+    sort(import_order) ==sort(c("dependencies", "main_package", "extensions"))
   )
-  if(!isTRUE(check_loadorder)) {
+  if(!isTRUE(check_import_order)) {
     return(FALSE)
   }