+++ R/import_misc.R
@@ -5,7 +5,7 @@
 #' The \code{import_LL()} function
 #' places specific functions from a package in the current environment,
 #' and also locks (see \link[base]{lockBinding}) the specified functions to prevent modification. \cr
-#' The primary use-case for this function is for loading functions inside a local environment,
+#' The primary use-case for this function is for exposing functions inside a local environment,
 #' like the environment within a function. \cr
 #' \cr
 #' The \code{import_int()} function
@@ -61,7 +61,7 @@
 #' The specified functions will be placed in the current environment
 #' (like the global environment, or the environment within a function),
 #' and locked. \cr
-#' To "unload" or overwrite the functions, simply remove them; i.e.: \cr
+#' To unexpose or overwrite the functions, simply remove them; i.e.: \cr
 #' \code{rm(list=c("some_function1", "some_function2")}). \cr
 #' \cr
 #' For \code{import_int()}: \cr